requirements.txt
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
boxmot==17.0.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
contourpy==1.3.0
coverage==7.2.7
cycler==0.12.1
deep-sort-realtime==1.3.2
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.128.8
ffmpy==1.0.0
filelock==3.19.1
filterpy==1.4.5
fonttools==4.60.2
fsspec==2025.10.0
ftfy==6.3.1
gdown==5.2.2
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
lapx==0.9.4
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
motmetrics==1.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.11.5
packaging==26.0
pandas==2.3.3
pillow==10.4.0
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.2
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
PySocks==1.7.1
pytest==8.3.3
pytest-cov==4.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.5
rich==15.0.0
ruff==0.6.9
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
tomlkit==0.12.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.66.5
transformers==4.57.6
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.39.0
wcwidth==0.6.0
websockets==12.0
xmltodict==1.0.4
yacs==0.1.8
zipp==3.23.1
